+++ lib/src/ast/selector.dart
@@ -5,6 +5,7 @@
 import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 
+import '../deprecation.dart';
 import '../evaluation_context.dart';
 import '../exception.dart';
 import '../visitor/any_selector.dart';
@@ -88,13 +89,13 @@ abstract class Selector implements AstNode {
   /// throw a [SassException] in Dart Sass 2.0.0.
   void assertNotBogus({String? name}) {
     if (!isBogus) return;
-    warn(
+    warnForDeprecation(
         (name == null ? '' : '\$$name: ') +
             '$this is not valid CSS.\n'
                 'This will be an error in Dart Sass 2.0.0.\n'
                 '\n'
                 'More info: https://sass-lang.com/d/bogus-combinators',
-        deprecation: true);
+        Deprecation.bogusCombinators);
   }
 
   /// Calls the appropriate visit method on [visitor].