+++ lib/src/executable/compile_stylesheet.dart
@@ -72,7 +72,9 @@ Future<void> compileStylesheet(ExecutableOptions options, StylesheetGraph graph,
               quietDeps: options.quietDeps,
               verbose: options.verbose,
               sourceMap: options.emitSourceMap,
-              charset: options.charset)
+              charset: options.charset,
+              fatalDeprecations: options.fatalDeprecations,
+              futureDeprecations: options.futureDeprecations)
           : await compileAsync(source,
               syntax: syntax,
               logger: options.logger,
@@ -81,7 +83,9 @@ Future<void> compileStylesheet(ExecutableOptions options, StylesheetGraph graph,
               quietDeps: options.quietDeps,
               verbose: options.verbose,
               sourceMap: options.emitSourceMap,
-              charset: options.charset);
+              charset: options.charset,
+              fatalDeprecations: options.fatalDeprecations,
+              futureDeprecations: options.futureDeprecations);
     } else {
       result = source == null
           ? compileString(await readStdin(),
@@ -93,7 +97,9 @@ Future<void> compileStylesheet(ExecutableOptions options, StylesheetGraph graph,
               quietDeps: options.quietDeps,
               verbose: options.verbose,
               sourceMap: options.emitSourceMap,
-              charset: options.charset)
+              charset: options.charset,
+              fatalDeprecations: options.fatalDeprecations,
+              futureDeprecations: options.futureDeprecations)
           : compile(source,
               syntax: syntax,
               logger: options.logger,
@@ -102,7 +108,9 @@ Future<void> compileStylesheet(ExecutableOptions options, StylesheetGraph graph,
               quietDeps: options.quietDeps,
               verbose: options.verbose,
               sourceMap: options.emitSourceMap,
-              charset: options.charset);
+              charset: options.charset,
+              fatalDeprecations: options.fatalDeprecations,
+              futureDeprecations: options.futureDeprecations);
     }
   } on SassException catch (error) {
     if (options.emitErrorCss) {