+++ lib/src/logger.dart
@@ -2,9 +2,12 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 import 'package:stack_trace/stack_trace.dart';
 
+import 'deprecation.dart';
+import 'logger/deprecation_handling.dart';
 import 'logger/stderr.dart';
 
 /// An interface for loggers that print messages produced by Sass stylesheets.
@@ -32,6 +35,22 @@ abstract class Logger {
 
   /// Emits a debugging message associated with the given [span].
   void debug(String message, SourceSpan span);
+}
+
+/// An extension to add a `warnForDeprecation` method to loggers without
+/// making a breaking API change.
+@internal
+extension WarnForDeprecation on Logger {
+  /// Emits a deprecation warning for [deprecation] with the given [message].
+  void warnForDeprecation(Deprecation deprecation, String message,
+      {FileSpan? span, Trace? trace}) {
+    var self = this;
+    if (self is DeprecationHandlingLogger) {
+      self.warnForDeprecation(deprecation, message, span: span, trace: trace);
+    } else if (!deprecation.isFuture) {
+      warn(message, span: span, trace: trace, deprecation: true);
+    }
+  }
 }
 
 /// A logger that emits no messages.