+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 8945d2e2978c178b096a4bbcd3857572ec5ab1e0
+// Checksum: 8a55729a9dc5dafe90954738907880052d930898
 //
 // ignore_for_file: unused_import
 
@@ -32,6 +32,7 @@ import '../callable.dart';
 import '../color_names.dart';
 import '../configuration.dart';
 import '../configured_value.dart';
+import '../deprecation.dart';
 import '../evaluation_context.dart';
 import '../exception.dart';
 import '../extend/extension_store.dart';
@@ -454,12 +455,12 @@ class _EvaluateVisitor
                     callableNode.span));
 
         if (function is SassString) {
-          warn(
+          warnForDeprecation(
               "Passing a string to call() is deprecated and will be illegal in "
               "Dart Sass 2.0.0.\n"
               "\n"
               "Recommendation: call(get-function($function))",
-              deprecation: true);
+              Deprecation.callString);
 
           var callableNode = _callableNode!;
           var expression =
@@ -1209,7 +1210,7 @@ class _EvaluateVisitor
                   'More info: https://sass-lang.com/d/bogus-combinators',
           MultiSpan(complex.span.trimRight(), 'invalid selector',
               {node.span: '@extend rule'}),
-          deprecation: true);
+          Deprecation.bogusCombinators);
     }
 
     var tuple = _performInterpolationWithMap(node.selector, warnForColor: true);
@@ -1915,7 +1916,7 @@ class _EvaluateVisitor
               '\n'
               'More info: https://sass-lang.com/d/bogus-combinators',
               complex.span.trimRight(),
-              deprecation: true);
+              Deprecation.bogusCombinators);
         } else if (complex.leadingCombinators.isNotEmpty) {
           _warn(
               'The selector "${complex.toString().trim()}" is invalid CSS.\n'
@@ -1923,7 +1924,7 @@ class _EvaluateVisitor
               '\n'
               'More info: https://sass-lang.com/d/bogus-combinators',
               complex.span.trimRight(),
-              deprecation: true);
+              Deprecation.bogusCombinators);
         } else {
           _warn(
               'The selector "${complex.toString().trim()}" is only valid for '
@@ -1942,7 +1943,7 @@ class _EvaluateVisitor
                         ? '\n(try converting to a //-style comment)'
                         : '')
               }),
-              deprecation: true);
+              Deprecation.bogusCombinators);
         }
       }
     }
@@ -2071,7 +2072,7 @@ class _EvaluateVisitor
                   "Recommendation: add `${node.originalName}: null` at the "
                   "stylesheet root.",
           node.span,
-          deprecation: true);
+          Deprecation.newGlobal);
     }
 
     var value = _withoutSlash(node.expression.accept(this), node.expression);
@@ -2206,7 +2207,7 @@ class _EvaluateVisitor
                   "More info and automated migrator: "
                   "https://sass-lang.com/d/slash-div",
                   node.span,
-                  deprecation: true);
+                  Deprecation.slashDiv);
             }
 
             return result;
@@ -3367,7 +3368,7 @@ class _EvaluateVisitor
           "More info and automated migrator: "
           "https://sass-lang.com/d/slash-div",
           nodeForSpan.span,
-          deprecation: true);
+          Deprecation.slashDiv);
     }
 
     return value.withoutSlash();
@@ -3390,15 +3391,20 @@ class _EvaluateVisitor
   }
 
   /// Emits a warning with the given [message] about the given [span].
-  void _warn(String message, FileSpan span, {bool deprecation = false}) {
+  void _warn(String message, FileSpan span, [Deprecation? deprecation]) {
     if (_quietDeps &&
         (_inDependency || (_currentCallable?.inDependency ?? false))) {
       return;
     }
 
     if (!_warningsEmitted.add(Tuple2(message, span))) return;
-    _logger.warn(message,
-        span: span, trace: _stackTrace(span), deprecation: deprecation);
+    var trace = _stackTrace(span);
+    if (deprecation == null) {
+      _logger.warn(message, span: span, trace: trace);
+    } else {
+      _logger.warnForDeprecation(deprecation, message,
+          span: span, trace: trace);
+    }
   }
 
   /// Returns a [SassRuntimeException] with the given [message].
@@ -3554,13 +3560,13 @@ class _EvaluationContext implements EvaluationContext {
     throw StateError("No Sass callable is currently being evaluated.");
   }
 
-  void warn(String message, {bool deprecation = false}) {
+  void warn(String message, [Deprecation? deprecation]) {
     _visitor._warn(
         message,
         _visitor._importSpan ??
             _visitor._callableNode?.span ??
             _defaultWarnNodeWithSpan.span,
-        deprecation: deprecation);
+        deprecation);
   }
 }